fix tagging tests
