+++ testing/tests/test.rs
@@ -1,4 +1,3 @@
-#![feature(specialization)]
 #![cfg_attr(feature = "clippy", feature(plugin))]
 #![cfg_attr(feature = "clippy", plugin(clippy))]
