+++ testing/tests/test_de.rs
@@ -814,42 +814,3 @@ declare_error_tests! {
         Error::DuplicateField("a"),
     }
 }
-
-#[test]
-fn tagging() {
-    #[derive(Debug, PartialEq, Eq)]
-    struct TaggedValue(String);
-
-    impl Deserialize for TaggedValue {
-        fn deserialize<D>(deserializer: &mut D) -> Result<TaggedValue, D::Error>
-            where D: serde::Deserializer,
-        {
-            deserializer.deserialize_tag()
-        }
-    }
-    assert_de_tokens(
-        &TaggedValue("cake".to_string()),
-        &[
-            Token::Tag,
-            Token::I32(42),
-            Token::String("cake".to_owned()),
-        ],
-    );
-
-    trait DeserializeTag<S: Deserialize>: serde::Deserializer {
-        fn deserialize_tag(&mut self) -> Result<S, <Self as serde::Deserializer>::Error>;
-    }
-
-    impl<I: Iterator<Item=Token<'static>>> DeserializeTag<TaggedValue> for Deserializer<I> {
-        fn deserialize_tag(&mut self) -> Result<TaggedValue, Self::Error> {
-            use serde::Deserializer;
-            Ok(TaggedValue(try!(self.deserialize_tagged_value())))
-        }
-    }
-
-    impl<S: serde::Deserializer> DeserializeTag<TaggedValue> for S {
-        default fn deserialize_tag(&mut self) -> Result<TaggedValue, S::Error> {
-            unimplemented!()
-        }
-    }
-}
