+++ testing/tests/test_ser.rs
@@ -376,41 +376,3 @@ fn test_cannot_serialize_paths() {
         &[],
         Error::InvalidValue("Path contains invalid UTF-8 characters".to_owned()));
 }
-
-#[test]
-fn tagging() {
-    struct TaggedValue(String);
-
-    trait SerializeTag<S: self::serde::Serialize>: self::serde::Serializer {
-        fn serialize_tag(&mut self, value: &S) -> Result<(), <Self as self::serde::Serializer>::Error>;
-    }
-
-    impl<S: self::serde::Serializer> SerializeTag<TaggedValue> for S {
-        default fn serialize_tag(&mut self, _value: &TaggedValue) -> Result<(), S::Error> {
-            unimplemented!()
-        }
-    }
-
-    impl<'a, I: Iterator<Item=&'a Token<'a>>> SerializeTag<TaggedValue> for Serializer<'a, I> {
-        fn serialize_tag(&mut self, value: &TaggedValue) -> Result<(), Self::Error> {
-            use self::serde::Serializer;
-            self.serialize_tagged_value(42, &value.0)
-        }
-    }
-
-    impl self::serde::Serialize for TaggedValue {
-        fn serialize<S>(&self, serializer: &mut S) -> Result<(), S::Error>
-            where S: self::serde::Serializer,
-        {
-            serializer.serialize_tag(self)
-        }
-    }
-    assert_ser_tokens(
-        &TaggedValue("cake".to_string()),
-        &[
-            Token::Tag,
-            Token::I32(42),
-            Token::Str("cake"),
-        ],
-    );
-}
